+++ src/util.rs
@@ -128,41 +128,22 @@ impl SegmentedJumpInitialVelocityCalculator {
     }
 }
 
-/// Represent a model's orientation in order to derive rotation instruction.
-pub struct ProjectionPlaneForRotation {
-    /// The direction where the character is facing in the model.
-    pub forward: Vector3,
-    /// The direction in the model that is the character's left.
-    pub sideways: Vector3,
-}
-
-impl ProjectionPlaneForRotation {
-    /// A projection plane for a model by "up" direction and "forward" direction (where the
-    /// character is facing)
-    pub fn from_up_and_fowrard(up: Direction3d, forward: Vector3) -> Self {
-        Self {
-            forward,
-            sideways: up.adjust_precision().cross(forward),
-        }
-    }
-
-    /// Project a 3D vector to this plane as a 2D vector using a coordinate system where
-    /// [`forward`](Self::forward) is the X axis.
-    pub fn project_and_normalize(&self, vector: Vector3) -> Vector2 {
-        Vector2::new(vector.dot(self.forward), vector.dot(self.sideways)).normalize_or_zero()
-    }
-
-    /// Assuming the character is currently facing `current_forward`, create a rotation on the
-    /// plane that will cause it to face `desired_forward`.
-    pub fn rotation_to_set_forward(
-        &self,
-        current_forward: Vector3,
-        desired_forward: Vector3,
-    ) -> Float {
-        let rotation_to_set_forward = Quaternion::from_rotation_arc_2d(
-            self.project_and_normalize(current_forward),
-            self.project_and_normalize(desired_forward),
-        );
-        rotation_to_set_forward.xyz().z
-    }
+/// Calculate the rotation around `around_axis` required to rotate the character from
+/// `current_forward` to `desired_forward`.
+pub fn rotation_arc_around_axis(
+    around_axis: Direction3d,
+    current_forward: Vector3,
+    desired_forward: Vector3,
+) -> Option<Float> {
+    let around_axis: Vector3 = around_axis.adjust_precision();
+    let rotation_plane_x = current_forward.reject_from(around_axis).try_normalize()?;
+    let rotation_plane_y = around_axis.cross(rotation_plane_x);
+    let desired_forward_in_plane_coords = Vector2::new(
+        rotation_plane_x.dot(desired_forward),
+        rotation_plane_y.dot(desired_forward),
+    )
+    .try_normalize()?;
+    let rotation_to_set_forward =
+        Quaternion::from_rotation_arc_2d(Vector2::X, desired_forward_in_plane_coords);
+    Some(rotation_to_set_forward.xyz().z)
 }